Changes int to size_t for patch to check_name for pedantry.

+++ 2bwm.c
@@ -326,7 +326,7 @@ void check_name(struct client *client)
         if (NULL!=reply) free(reply);
         return;
     }
-    int reply_len = xcb_get_property_value_length(reply);
+    size_t reply_len = xcb_get_property_value_length(reply);
     char *wm_name_window = malloc(sizeof(char) * (reply_len + 1));
     memcpy(wm_name_window, xcb_get_property_value(reply), reply_len);
     wm_name_window[reply_len] = '\0';
